feat: create player view
