+++ src/config/appSettingTypes.ts
@@ -1,1 +1,2 @@
 export const MOCK_SETTING_KEY = 'mock_setting';
+export const TEXT_RESOURCE_KEY = 'text_resource_setting';
